feat(eslint-plugin): replace barrel-import with gozd/no-define-expose
intent(eslint-plugin): repurpose the local plugin as the home for gozd-specific rules now that barrel-import lives upstream; first rule blocks defineExpose to enforce the props / composable boundary (issue #291)
decision(no-define-expose): match by CallExpression callee Identifier name only — Vue's compiler macro is never imported, so name match is sufficient and avoids needing vue-eslint-parser awareness
rejected(no-define-expose): selector string `CallExpression[callee.name='defineExpose']` — explicit visitor reads better and matches the project's existing rule style
constraint(eslint-plugin): noInlineConfig is on, so existing offenders (SidebarMenu.vue, XtermTerminal.vue) cannot be suppressed inline and must be migrated to satisfy lint
