+++ apps/renderer/package.json
@@ -33,6 +33,7 @@
     "vue": "3.5.34"
   },
   "devDependencies": {
+    "@gozd/eslint-plugin": "workspace:*",
     "@iconify-json/lucide": "1.2.108",
     "@iconify/tailwind4": "1.2.3",
     "@miyaoka/eslint-plugin-barrel-import": "0.0.0",
